+++ CHANGES.rst
@@ -41,6 +41,8 @@ Version 3.2.0
 -   Modifying ``Response.mimetype_params`` when ``mimetype`` is not set does
     nothing instead of setting an invalid header.
 -   ``WWWAuthenticate.type`` is empty by default instead of ``basic``.
+-   Fix setting ``WWWAuthenticate`` attributes. Get/set/delete on arbitrary
+    ``wa.key`` and ``wa[key]`` is deprecated, use ``wa.parameters`` instead.
 -   The ``Request.parameter_storage_class``, ``dict_storage_class`` and
     ``list_storage_class`` attributes, and the ``cls`` parameter to
     ``parse_cookie``, ``parse_form_data``, and ``FormDataParser``, are